Update plugin.py
Removed `_create_unverified_https_context` for legacy Python that doesn't verify HTTPS certificates by default. It is not neccessary in this case. 

Older versions of Python do not support and need to ignore SSL certificates. In this case, older versions of Python (such as 2.7.2 in the DMM image) cause error messages about the non-existent `_create_unverified_https_context attribute`.

I am using the `urllib2` library to download files from HTTPS servers, but without an SSL connection, which is often problematic and complicated.